refactor: remove unused dependencies and components, and restructure calendar tiles
- Removed @chakra-ui/react and @emotion/react from dependencies.
- Deleted CreateEventAction component as it was not in use.
- Replaced MonthlyTile with CalendarTile in the calendar component.
- Deleted MonthlyTile component and moved its logic to CalendarTile.
- Introduced placerEvenementsEnColonne function to handle event positioning logic.
- Updated constants for weekdays and months in a dedicated utils file.
- Adjusted PlanningPage to calculate year in milliseconds dynamically.
