--- src/app/planning/components/actions/create-event/index.tsx
@@ -1,9 +1,0 @@
-export default function CreateEventAction() {
-  return (
-    <div className="create-event-box">
-      {/* <h2>Créer un nouvel événement</h2>
-
-      <form className="create-event-form"></form> */}
-    </div>
-  );
-}
