+++ src/app/planning/components/calendar/calendar.tsx
@@ -1,7 +1,6 @@
 "use client";
 import React, { useState, useEffect } from "react";
-import MonthlyTile from "./tiles/monthly-tile";
-import CreateEventAction from "../actions/create-event";
+import CalendarTile from "./tiles/calendar-tiles";
 import {
   Alert,
   Badge,
@@ -24,6 +23,7 @@ import {
   Icon,
   SearchBar,
 } from "@/components/ui";
+import { MINI_WEEKDAYS, MONTHS, WEEKDAYS } from "./utils/constant";
 
 export type EvenementBuiltInt = {
   id: number;
@@ -53,30 +53,7 @@ interface EventCalendarProps {
   onEventClick?: (event: EvenementBuiltInt) => void;
 }
 
-const WEEKDAYS = [
-  { full: "Lundi", short: "Lun" },
-  { full: "Mardi", short: "Mar" },
-  { full: "Mercredi", short: "Mer" },
-  { full: "Jeudi", short: "Jeu" },
-  { full: "Vendredi", short: "Ven" },
-  { full: "Samedi", short: "Sam" },
-  { full: "Dimanche", short: "Dim" },
-];
-const MINI_WEEKDAYS = ["Lu", "Ma", "Me", "Je", "Ve", "Sa", "Di"];
-const MONTHS = [
-  "Janvier",
-  "Fevrier",
-  "Mars",
-  "Avril",
-  "Mai",
-  "Juin",
-  "Juillet",
-  "Aout",
-  "Septembre",
-  "Octobre",
-  "Novembre",
-  "Decembre",
-];
+
 
 const Calendar: React.FC<EventCalendarProps> = ({ events, onEventClick }: EventCalendarProps) => {
   const [currentDate, setCurrentDate] = useState(new Date());
@@ -433,7 +410,6 @@ const Calendar: React.FC<EventCalendarProps> = ({ events, onEventClick }: EventC
       </div>
 
       <div className="calendar">
-        <CreateEventAction />
         <div className="calendar-content">
           {viewType === "weekly" && (
             <div className="time-slots" ref={ref}>
@@ -468,7 +444,7 @@ const Calendar: React.FC<EventCalendarProps> = ({ events, onEventClick }: EventC
                 ))}
 
               {calendarDays.map((calDay, index) => (
-                <MonthlyTile
+                <CalendarTile
                   key={index}
                   calDay={calDay}
                   index={index}
