+++ src/app/planning/page.tsx
@@ -1,6 +1,7 @@
 import Calendar from "./components/calendar/calendar";
 
-const yearInMS = 31536000000;
+const yearInMS = new Date(2027, 0, 1).getTime() - new Date(2026, 0, 1).getTime();
+
 
 export default function PlanningPage() {
   return (
